Refactored functions names
