+++ src/index.test.ts
@@ -30,9 +30,9 @@ import {
   getAccessGrant,
   getAccessGrantAll,
   getAccessApiEndpoint,
-  getGrantFromRedirectUrl,
+  getAccessGrantFromRedirectUrl,
   getAccessManagementUi,
-  getRequestFromRedirectUrl,
+  getAccessRequestFromRedirectUrl,
   isValidAccessGrant,
   issueAccessRequest,
   redirectToAccessManagementUi,
@@ -63,9 +63,9 @@ describe("Index exports", () => {
     expect(getAccessGrant).toBeDefined();
     expect(getAccessGrantAll).toBeDefined();
     expect(getAccessApiEndpoint).toBeDefined();
-    expect(getGrantFromRedirectUrl).toBeDefined();
+    expect(getAccessGrantFromRedirectUrl).toBeDefined();
     expect(getAccessManagementUi).toBeDefined();
-    expect(getRequestFromRedirectUrl).toBeDefined();
+    expect(getAccessRequestFromRedirectUrl).toBeDefined();
     expect(isValidAccessGrant).toBeDefined();
     expect(issueAccessRequest).toBeDefined();
     expect(redirectToAccessManagementUi).toBeDefined();
