+++ src/request/getAccessGrantFromRedirectUrl.test.ts
@@ -22,9 +22,9 @@
 // eslint-disable-next-line no-shadow
 import { describe, it, jest, expect } from "@jest/globals";
 import { getVerifiableCredential } from "@inrupt/solid-client-vc";
-import { getRequestFromRedirectUrl } from "./getRequestFromRedirectUrl";
-import { mockAccessGrantVc } from "./approveAccessRequest.mock";
+import { getAccessGrantFromRedirectUrl } from "./getAccessGrantFromRedirectUrl";
 import { getSessionFetch } from "../util/getSessionFetch";
+import { mockAccessGrantVc } from "../manage/approveAccessRequest.mock";
 
 jest.mock("../util/getSessionFetch");
 jest.mock("@inrupt/solid-client-vc", () => {
@@ -38,29 +38,11 @@ jest.mock("@inrupt/solid-client-vc", () => {
   };
 });
 
-describe("getRequestFromRedirectUrl", () => {
-  it("throws if the requestVcUrl query parameter is missing", async () => {
-    const redirectUrl = new URL("https://redirect.url");
-    redirectUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
-    );
-
-    await expect(getRequestFromRedirectUrl(redirectUrl.href)).rejects.toThrow(
-      /https:\/\/redirect.url.*requestVcUrl/
-    );
-  });
-
-  it("throws if the redirectUrl query parameter is missing", async () => {
-    const redirectUrl = new URL("https://redirect.url");
-    redirectUrl.searchParams.append(
-      "requestVcUrl",
-      encodeURI("https://some.vc")
-    );
-
-    await expect(getRequestFromRedirectUrl(redirectUrl.href)).rejects.toThrow(
-      /https:\/\/redirect.url.*redirectUrl/
-    );
+describe("getAccessGrantFromRedirectUrl", () => {
+  it("throws if the accessGrant query parameter is missing", async () => {
+    await expect(
+      getAccessGrantFromRedirectUrl("https://redirect.url")
+    ).rejects.toThrow(/https:\/\/redirect.url.*accessGrant/);
   });
 
   it("uses the default fetch if none is provided", async () => {
@@ -77,17 +59,13 @@ describe("getRequestFromRedirectUrl", () => {
     }>;
     vcModule.getVerifiableCredential.mockResolvedValueOnce(mockAccessGrantVc());
 
-    const redirectedToUrl = new URL("https://redirect.url");
-    redirectedToUrl.searchParams.append(
-      "requestVcUrl",
+    const redirectUrl = new URL("https://redirect.url");
+    redirectUrl.searchParams.set(
+      "accessGrantUrl",
       encodeURI("https://some.vc")
     );
-    redirectedToUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
-    );
 
-    await getRequestFromRedirectUrl(redirectedToUrl.href, {
+    await getAccessGrantFromRedirectUrl(redirectUrl.href, {
       fetch: mockedFetch,
     });
     expect(vcModule.getVerifiableCredential).toHaveBeenCalledWith(
@@ -106,18 +84,14 @@ describe("getRequestFromRedirectUrl", () => {
     }>;
     vcModule.getVerifiableCredential.mockResolvedValueOnce(mockAccessGrantVc());
 
-    const redirectedToUrl = new URL("https://redirect.url");
-    redirectedToUrl.searchParams.append(
-      "requestVcUrl",
+    const redirectUrl = new URL("https://redirect.url");
+    redirectUrl.searchParams.set(
+      "accessGrantUrl",
       encodeURI("https://some.vc")
-    );
-    redirectedToUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
     );
 
     const mockedFetch = jest.fn(fetch);
-    await getRequestFromRedirectUrl(redirectedToUrl.href, {
+    await getAccessGrantFromRedirectUrl(redirectUrl.href, {
       fetch: mockedFetch,
     });
     expect(vcModule.getVerifiableCredential).toHaveBeenCalledWith(
@@ -128,7 +102,7 @@ describe("getRequestFromRedirectUrl", () => {
     );
   });
 
-  it("returns the fetched VC and the redirect URL", async () => {
+  it("returns the fetched VC", async () => {
     const vcModule = jest.requireMock(
       "@inrupt/solid-client-vc"
     ) as jest.Mocked<{
@@ -136,57 +110,43 @@ describe("getRequestFromRedirectUrl", () => {
     }>;
     vcModule.getVerifiableCredential.mockResolvedValueOnce(mockAccessGrantVc());
 
-    const redirectedToUrl = new URL("https://redirect.url");
-    redirectedToUrl.searchParams.append(
-      "requestVcUrl",
+    const redirectUrl = new URL("https://redirect.url");
+    redirectUrl.searchParams.set(
+      "accessGrantUrl",
       encodeURI("https://some.vc")
     );
-    redirectedToUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
-    );
 
-    const { accessRequest, redirectUrl } = await getRequestFromRedirectUrl(
-      redirectedToUrl.href
-    );
-    expect(accessRequest).toStrictEqual(mockAccessGrantVc());
-    expect(redirectUrl).toBe("https://requestor.redirect.url");
+    const fetchedVc = await getAccessGrantFromRedirectUrl(redirectUrl.href);
+    expect(fetchedVc).toStrictEqual(mockAccessGrantVc());
   });
 
-  it("supports the legacy approach of providing the VC as a value", async () => {
+  it("supports the legacy approach where the VC is provided by value", async () => {
     const vcModule = jest.requireMock(
       "@inrupt/solid-client-vc"
     ) as jest.Mocked<{
       getVerifiableCredential: typeof getVerifiableCredential;
     }>;
-
     const redirectUrl = new URL("https://redirect.url");
-    redirectUrl.searchParams.append(
-      "requestVc",
-      encodeURIComponent(JSON.stringify(mockAccessGrantVc()))
-    );
-    redirectUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
+    redirectUrl.searchParams.set(
+      "accessGrant",
+      btoa(JSON.stringify(mockAccessGrantVc()))
     );
 
-    const { accessRequest } = await getRequestFromRedirectUrl(redirectUrl.href);
-    expect(accessRequest).toStrictEqual(mockAccessGrantVc());
-    // When the VC is passed as a value, nothing needs to be dereferenced.
+    const grantVc = await getAccessGrantFromRedirectUrl(redirectUrl.href);
+    expect(grantVc).toStrictEqual(mockAccessGrantVc());
+    // If the VC is provided as a value, no dereferencing should happen.
     expect(vcModule.getVerifiableCredential).not.toHaveBeenCalled();
   });
 
   it("throws if the legacy provided value is not a VC", async () => {
     const redirectUrl = new URL("https://redirect.url");
-    redirectUrl.searchParams.append(
-      "requestVc",
-      encodeURIComponent(JSON.stringify({ someJson: "but not a VC" }))
-    );
-    redirectUrl.searchParams.append(
-      "redirectUrl",
-      encodeURI("https://requestor.redirect.url")
+    redirectUrl.searchParams.set(
+      "accessGrant",
+      btoa(JSON.stringify({ someJson: "but not a VC" }))
     );
 
-    await expect(getRequestFromRedirectUrl(redirectUrl.href)).rejects.toThrow();
+    await expect(
+      getAccessGrantFromRedirectUrl(redirectUrl.href)
+    ).rejects.toThrow();
   });
 });
